requirements.txt
asgiref==3.9.1
charset-normalizer==3.4.2
Django==5.2.4
djangorestframework==3.16.0
pillow==11.3.0
psycopg2-binary==2.9.10
reportlab==4.4.3
sqlparse==0.5.3
tzdata==2025.2
